FIX: ajuste do tipo das chaves primarias de Bigint para serial

+++ Atividade central da disciplina/Criacao de tabelas/criar_tabelas.sql
@@ -1,5 +1,5 @@
 CREATE TABLE loja(
-    cod_loja bigint not null,
+    cod_loja serial not null,
     nome text not null,
     telefone varchar(255) not null,
     cep varchar(255),
@@ -13,41 +13,41 @@ CREATE TABLE loja(
 );
 
 CREATE TABLE fornecedor(
-    cod_fornecedor bigint not null,
+    cod_fornecedor serial not null,
     nome text not null,
     telefone varchar(255),
     constraint pk_fornecedor primary key(cod_fornecedor) 
 );
 
 CREATE TABLE produto(
-    cod_produto bigint not null,
+    cod_produto serial not null,
     nome text not null,
     categoria varchar(255),
     constraint pk_produto primary key(cod_produto)
 );
 
 CREATE TABLE fornecedor(
-    cod_fornecedor bigint not null,
+    cod_fornecedor serial not null,
     nome text not null,
     telefone varchar(20),
     constraint pk_fornecedor primary key(cod_fornecedor)
 );
 CREATE TABLE cliente(
-    cod_cliente bigint not null,
+    cod_cliente serial not null,
     nome text not null,
     telefone varchar(20),
     constraint pk_cliente primary key(cod_cliente)
 );
 
 CREATE TABLE cargo(
-    cod_cargo bigint not null,
+    cod_cargo serial  not null,
     nome varchar(255) not null,
     descricao text,
     constraint pk_cargo primary key(cod_cargo)
 );
 
 CREATE TABLE funcionario(
-    cod_funcionario bigint not null,
+    cod_funcionario serial not null,
     nome text not null,
     telefone varchar(20),
     salario decimal(30,2),
